Fix a datafetcher bug that might be throwing all kinds of shit off

+++ datafetcher/src/github.com/MustWin/datafetcher/nfl/models/types.go
@@ -26,7 +26,7 @@ type Market struct {
 	Id            int
 	Name          string
 	SportId       int `model:"key"`
-	TotalBets     int
+	TotalBets     float64
 	ShadowBets    int
 	ShadowBetRate float64
 	PublishedAt   time.Time
